gql for getWeek + update input to getWeek + new helper file
