+++ rest/v0/schedule.js
@@ -3,7 +3,7 @@ var router = express.Router();
 var {createErrorJSON} = require("../../helpers/errors.helper")
 
 var {callWebSocAPI} = require('websoc-api');
-var {getWeek} = require('../../helpers/schedule.helper')
+var {getWeek} = require('../../helpers/week.helper')
 
 router.get("/soc", function (req, res, next) {
     callWebSocAPI(req.query).then((val) => {
@@ -14,10 +14,10 @@ router.get("/soc", function (req, res, next) {
 })
 
 router.get("/week", function (req, res, next) {
-    getWeek().then((val) => {
+    getWeek(req.query.year, req.query.month, req.query.day).then((val) => {
         res.json(val)
     }).catch((err) => {
-        res.status(400).json(createErrorJSON(400, "Bad Request: Invalid parameter", "Unable to get week"));
+        res.status(400).json(createErrorJSON(400, "Bad Request: Invalid parameter", "Invalid date"));
     })
 })
